pyproject.toml
[tool.poetry]
name = "mp3-to-mp4"
version = "0.1.1"
description = "A CLI application for converting audio files (.mp3, .wav, .flac, .aiff, etc) to mp4 file suitable for upload to social media sites like YouTube, Vimeo, etc."
authors = ["Henry Oberholtzer <henryoberholtzer@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
mp3-to-mp4 = "mp3_to_mp4.__main__:main"

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.9.0"}
moviepy = "^1.0.3"
pytest = "^8.1.1"
pytest-mock = "^3.12.0"
tinytag = "^1.10.1"
pytest-cov = "^4.1.0"
pydub = "^0.25.1"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
